+++ src/common.ts
@@ -43,12 +43,6 @@ export function binarySearch<T>(array: T[], key: T, comparator: (op1: T, op2: T)
     return -(low + 1);
 }
 
-// export const EMPTY_ELEMENTS: string[] = ['area', 'base', 'br', 'col', 'embed', 'hr', 'img', 'input', 'keygen', 'link', 'menuitem', 'meta', 'param', 'source', 'track', 'wbr'];
-
-// export function isEmptyElement(e: string): boolean {
-//     return !!e && binarySearch(EMPTY_ELEMENTS, e.toLowerCase(), (s1: string, s2: string) => s1.localeCompare(s2)) >= 0;
-// }
-
 //
 
 export function buildMap<T>(obj: {[name: string]: T}) {